+++ RELEASE_NOTES.md
@@ -34,6 +34,9 @@ Logtalk compiler and runtime
 
 * IMPROVED: More efficient compiler checking for parameter variables.
 
+* FIXED: Using a overriding meta-predicate template for predicates defined in
+`user` listed in `uses/2` directives.
+
 * FIXED: Bug in the `logtalk_make/0-1` predicates when reloading a main file
 due to an updated included file when the `clean` flag is turned off and the
 `reload` file is set tp `changed` where the main file would not be recompiled